* 调整部分动画相关接口，使之现代化（todo: editor 脚本需要跟着一起改）；优化渲染图形系统对于摄像机和实体的排序实现
* 调整部分图形接口的实现以确保行为正确

+++ src/jeecs_graphic_api_basic.cpp
@@ -1726,7 +1726,7 @@ void jegl_draw_vertex(jegl_resource* vert)
 
 void jegl_rend_to_framebuffer(
     jegl_resource* framebuffer,
-    const size_t(*viewport_xywh)[4],
+    const int32_t(*viewport_xywh)[4],
     const float (*clear_color_rgba)[4],
     const float* clear_depth)
 {
